✨ feat(stores): ThreadRunStore::checkpoint_append version-guarded default
ADR-0042 A: append delta + run atomically, guarded by committed count.
Extract storage.rs inline tests to storage/tests.rs for the length gate.
