+++ crates/awaken-stores/tests/memory_thread_store_conformance.rs
@@ -19,6 +19,9 @@ conformance_test!(latest_run_returns_most_recent);
 conformance_test!(checkpoint_overwrites_messages);
 conformance_test!(load_thread_reflects_checkpoint);
 conformance_test!(append_message_records_assigns_seq);
+conformance_test!(checkpoint_append_assigns_version);
+conformance_test!(checkpoint_append_unconditional_appends);
+conformance_test!(checkpoint_append_rejects_stale_version);
 conformance_test!(list_threads_query_filters_lineage);
 conformance_test!(list_threads_query_filters_root_threads);
 conformance_test!(checkpoint_rejects_missing_parent_thread);
